requirements.txt
# VacationManager v5.5 - Python Dependencies

# Backend Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
aiosqlite>=0.19.0

# Desktop UI
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0

# Document Generation
docxtpl>=0.16.0
python-docx>=1.1.0

# Ukrainian Grammar
pymorphy3
pymorphy3-dicts-uk

# Web Portal
jinja2>=3.1.0
python-multipart>=0.0.6

# Validation & Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dateutil>=2.8.0
structlog